requirements.txt
pillow
whitenoise
Django
djangorestframework
django-filter  # Filtering support
markdown       # Markdown support for the browsable API.
django-allauth
django-environ
django-npm
django-tinymce

